refactor(moderation): split ModerationController into Pages + Mutations
The single ModerationController had grown to 825 lines mixing 13
@GetMapping page handlers, ~25 @PostMapping/@DeleteMapping mutations,
chart pre-computation for the Activity tab, and ~20 request/response
DTOs. Split it along the read/write axis — the natural cohesion line
already implied by the @GetMapping vs @PostMapping/@DeleteMapping
annotations:

  - web/controller/moderation/ModerationPagesController.kt (~245
    lines): every page handler + the shared `renderSubPage` helper +
    the OAuth-gated `guildList` lobby. Wires `ActivityChartsService`
    for the Activity-tab pre-compute.
  - web/controller/moderation/ModerationMutationsController.kt (~420
    lines plus colocated DTOs): every POST/DELETE handler + the
    `notSignedInApi` helper.

Both share `@RequestMapping("/moderation")` — Spring tolerates
multiple controllers under one base, and the URL space stays
identical (no routing changes, no template changes, no JS changes).
The single ModerationController bean's responsibilities flow into
the two new beans; component scanning picks up both automatically.

Existing tests relocated to `web.controller.moderation` package and
renamed:
  - ModerationControllerTest → ModerationMutationsControllerTest
  - ModerationControllerWelcomeTest → ModerationMutationsControllerWelcomeTest

The mutations-controller constructor is simpler (one `ModerationWebService`
dependency instead of three), so the test setup drops the
`activityChartsService` and `discordClientId` mocks.
